requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.4.0
django-filter==24.2
django-simple-history==3.7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
inflection==0.5.1
packaging==24.0
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.7.0
